Fix Pr/56 Dataloader Refactor(#61)
* Dataloader Refactor
* fix docs
