+++ contk/dataloader/__init__.py
@@ -5,10 +5,10 @@ data and provide a fair dataset for every model. It also help you adapt
 your model from one dataset to other datasets.
 """
 
-from .dataloader import Dataloader
+from .dataloader import Dataloader, BasicLanguageGeneration
 from .single_turn_dialog import SingleTurnDialog, OpenSubtitles
 from .multi_turn_dialog import MultiTurnDialog, UbuntuCorpus
 from .language_generation import LanguageGeneration, MSCOCO
 
 __all__ = ['Dataloader', 'SingleTurnDialog', 'OpenSubtitles', 'MultiTurnDialog', 'UbuntuCorpus', \
-        'LanguageGeneration', 'MSCOCO']
+        'LanguageGeneration', 'MSCOCO', 'BasicLanguageGeneration']
